% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble_gene_set.R
\name{fetch_gene}
\alias{fetch_gene}
\title{Extract a gene sequence from a genbank file}
\usage{
fetch_gene(gene_name, gb_file, rename = TRUE, detect_dups = TRUE)
}
\arguments{
\item{gene_name}{Name of gene to extract. Must match genbank data exactly,
case-sensitive.}

\item{gb_file}{efetch object; full genbank record downloaded using reutils::efetch().
Typically a (sub)genome containing the gene of interest.}

\item{rename}{Logical; should the sequence be renamed as "accession-gene"?}

\item{detect_dups}{Logical; should duplicates be detected? If TRUE, a text
warning will be returned in the case that this gene has duplicate copies
in the genome; if false, all copies will be returned.}
}
\value{
object of class DNAbin
}
\description{
If duplicates are detected via `detect_dups` and `rename` is TRUE, duplicates
will be differentiated by appending a numeral to the name, e.g.
"accession-gene-1", "accession-gene-2", etc.
}
\details{
See https://github.com/gschofl/biofiles
}
\examples{
\dontrun{
Cystopteris_protrusa_plastome <- reutils::efetch("KP136830",
"nuccore", rettype = "gbwithparts", retmode = "text")
fetch_gene("rbcL", Cystopteris_protrusa_plastome)

Diplazium_striatum_plastome <- reutils::efetch("KY427346",
"nuccore", rettype = "gbwithparts", retmode = "text")

fetch_gene("psbA", Diplazium_striatum_plastome)
fetch_gene("psbA", Diplazium_striatum_plastome, detect_dups = FALSE)
fetch_gene("psbA", Diplazium_striatum_plastome, rename = FALSE, detect_dups = FALSE)
}

}
