% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble_gene_set.R
\name{fetch_gene_from_genome}
\alias{fetch_gene_from_genome}
\title{Extract a list of gene sequences from a genbank file}
\usage{
fetch_gene_from_genome(gene, accession, rename = TRUE,
  detect_dups = TRUE)
}
\arguments{
\item{gene}{Characte vector; gene names to extract.}

\item{accession}{String; GenBank accession number of (full or partial) genome
containing the gene(s) of interest.}

\item{rename}{Logical; should the sequence be renamed as "accession-gene"?}

\item{detect_dups}{Logical; should duplicates be detected? If TRUE, a text
warning will be returned in the case that this gene has duplicate copies
in the genome instead of the sequence; if false, all copies will be returned,
differentiated by appending numbers to the gene name.}
}
\value{
A list
}
\description{
Extract a list of gene sequences from a genbank file
}
\examples{
\dontrun{
# KY427346 is the GenBank accession no. for the Diplazium striatum plastome
# https://www.ncbi.nlm.nih.gov/nuccore/KY427346

genes_to_get <- c("rbcL", "matK", "psbA")

fetch_gene_from_genome(genes_to_get, "KY427346")
}

}
