% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble_gene_set.R
\name{fetch_gene_list}
\alias{fetch_gene_list}
\title{Extract a list of gene sequences from a genbank file}
\usage{
fetch_gene_list(gb_file, gene_vec, detect_dups = TRUE)
}
\arguments{
\item{gb_file}{efetch object; full genbank record downloaded using reutils::efetch().
Typically a (sub)genome containing the gene of interest.}

\item{gene_vec}{Vector of gene names to extract.}

\item{detect_dups}{Logical; should duplicates be detected? If TRUE, a text
warning will be returned in the case that this gene has duplicate copies
in the genome instead of the sequence; if false, all copies will be returned,
differentiated by appending numbers to the gene name.}
}
\value{
A list
}
\description{
Extract a list of gene sequences from a genbank file
}
\examples{
\dontrun{
Cystopteris_protrusa_plastome <- reutils::efetch("KP136830",
"nuccore", rettype = "gbwithparts", retmode = "text")
genes_to_get <- c("rbcL", "matK", "psbA")
fetch_gene_list(Cystopteris_protrusa_plastome, genes_to_get)

Diplazium_striatum_plastome <- reutils::efetch("KY427346",
"nuccore", rettype = "gbwithparts", retmode = "text")

fetch_gene_list(Diplazium_striatum_plastome, genes_to_get)
fetch_gene_list(Diplazium_striatum_plastome, genes_to_get, detect_dups = FALSE)
}

}
