% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_metadata.R
\name{fetch_metadata}
\alias{fetch_metadata}
\title{Fetch metadata from GenBank}
\usage{
fetch_metadata(query, chunk_size = 10, max_tries = 10,
  verbose = FALSE)
}
\arguments{
\item{query}{String used to query NCBI GenBank. For more about the NCBI
query format see
https://www.ncbi.nlm.nih.gov/books/NBK3837/#EntrezHelp.Entrez_Searching_Options}

\item{chunk_size}{Number of ids to use for each chunk. Changing this
doesn't tend to affect the results, but lower values have more accurate
progress bars.}

\item{max_tries}{Maximum number of times to attempt the loop.}

\item{verbose}{Logical; should information about number of loops attempted
be printed to the screen?}
}
\value{
Tibble of metadata resulting from Genbank query.
Columns include: \describe{
  \item{gi}{Genbank GI number}
  \item{accession}{Genbank accession number}
  \item{taxid}{Taxon ID (can use to query with taxize)}
  \item{title}{Sequence title}
  \item{slen}{Sequence length}
  \item{subname}{Misc. data (specimen, collection country, etc), separated
  by |}
  \item{subtype}{Column names of misc. data, separated by |}
  \item{species}{Species name}
}
}
\description{
Sequences downloaded from GenBank with \code{\link{fetch_sequences}} only include
the title and/or accession number. Use \code{fetch_metadata} to obtain other
useful metadata associated with the sequences.
}
\details{
\code{\link[rentrez]{entrez_search}} is used to obtain a vector of IDs from the
`query`, then \code{\link[rentrez]{entrez_summary}} is used to download metadata
from the IDs. However, \code{\link[rentrez]{entrez_summary}} will fail if too many
IDs are used as input (more than 200-300 or so). Therefore, \code{fetch_metadata}
splits the IDs into chunks (a list of vectors), and loops over the list.

Sometimes errors are encountered during the loop due to the API rejecting
the request, internet connectivity, etc. To avoid this, the loop repeats
until it finishes or the number of repeats reaches `max_tries`, upon
which it quits with an error.
}
\examples{
\dontrun{
fetch_metadata("rbcl[Gene] AND Crepidomanes[ORGN]")
}
}
