% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_metadata.R
\name{fetch_metadata_chunked}
\alias{fetch_metadata_chunked}
\title{Fetch metadata from GenBank sequences in chunks}
\usage{
fetch_metadata_chunked(id, col_select = c("gi", "caption", "taxid",
  "title", "slen", "subtype", "subname"), chunk_size = 100)
}
\arguments{
\item{id}{Character vector of IDs for GenBank records.
Output of rentrez::entrez_search().}

\item{col_select}{Vector of column names to return.}

\item{chunk_size}{Number of rows to use for each chunk.}
}
\value{
A list of two items, `results` and `error`.
}
\description{
fetch_metadata() queries the NCBI API online, so we should do this
in smaller chunks to avoid errors.
}
\examples{
\dontrun{
fetch_metadata_chunked(c("383212727", "383212725"))
}
}
