% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_sequences.R
\name{fetch_sequences}
\alias{fetch_sequences}
\title{Fetch DNA sequences from GenBank}
\usage{
fetch_sequences(query, simple_names = TRUE, chunk_size = 100,
  .pb = NULL, ...)
}
\arguments{
\item{query}{String used to query NCBI GenBank. For more about the NCBI
query format see
https://www.ncbi.nlm.nih.gov/books/NBK3837/#EntrezHelp.Entrez_Searching_Options}

\item{simple_names}{Logical; should the sequence names be simplified to
the GenBank accession number only?}

\item{chunk_size}{Number of ids to use for each chunk. Changing this
doesn't tend to affect the results, but lower values have more accurate
progress bars.}

\item{.pb}{Internal agument used for setting the progress bar; don't change this.}

\item{...}{Additional arguments, not used by this function but meant for enabling
tracking if this function is used as part of a \code{\link[drake]{drake_plan}}.}
}
\value{
List
}
\description{
Query GenBank using the same format as searches on the
\href{https://www.ncbi.nlm.nih.gov/nucleotide/}{NCBI nucleotide database} and
download the sequences directly into R.
}
\details{
\code{\link[rentrez]{entrez_search}} is used to obtain a vector of IDs from the
`query`, then downloads the corresponding DNA sequences
from the IDs. However, \code{\link[rentrez]{entrez_search}} will fail if too many
IDs are used as input (more than 200-300 or so). Therefore, \code{fetch_sequences}
splits the IDs into chunks (a list of vectors), and loops over the list.
}
\examples{
fetch_sequences("Crepidomanes minutum[ORGN] AND rbcl[Gene]")
}
